Move enum options models into nodegraph library
