+++ src/gui/models/nodeGraph/enumRegistry.cpp
@@ -3,7 +3,7 @@
 
 #include "enumRegistry.h"
 #include "data/structureFactors.h"
-#include "gui/models/enumOptionsModel.h"
+#include "gui/models/nodeGraph/enumOptionsModel.h"
 #include "math/windowFunction.h"
 #include "nodes/gr/gr.h"
 #include "nodes/md/md.h"
